engine/engine-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.2.0
numpy==1.26.0
psycopg2-binary==2.9.9
pydantic==2.6.0
python-dotenv==1.0.0
pytest==8.0.0
httpx==0.27.0
